pyproject.toml
[project]
name = "exasol-script-languages-package-management"
version = "0.1.0"
requires-python = ">=3.10,<3.15"
description = "This project provides a unified way of installing different packages from sources like coda, apt and PyPi"
authors = [
    {name = "Exasol AG", email = "opensource@exasol.com"},
]
readme = "README.md"
license = "MIT"
keywords = ['exasol', 'script-languages-package-management']
dynamic = ["dependencies"]

[tool.poetry]
requires-poetry = ">=2.1.0"
packages = [
    { include = "exasol" },
]

[poetry.urls]
repository = "https://github.com/exasol/script-languages-package-management"
homepage = "https://github.com/exasol/script-languages-package-management"

[tool.poetry.dependencies]
click = ">=8.3.0, <9.0.0"
pydantic="^2.10.2"
pyyaml = "^6.0.3"
types-pyyaml = "^6.0.12.20250915"

[tool.poetry.group.dev.dependencies]
exasol-toolbox = ">=1.6.0, <2.0.0"
pyinstaller = ">=6.15.0, <7.0.0"
types-pyinstaller = ">=6.15.0.20250822, <7.0.0.0"
docker = { version = ">=7.1.0,<8.0.0", markers = "sys_platform != 'win32'" }

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

testpaths = [
    "test"
]

[tool.coverage.run]
relative_files = true
source = [
    "exasol",
]

[tool.coverage.report]
fail_under = 0
ignore_errors = true

[tool.black]
line-length = 88
verbose = false
include = "\\.pyi?$"

[tool.isort]
profile = "black"
force_grid_wrap = 2

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = """
    R,C,
    duplicate-code,
    too-many-ancestors,
    broad-exception-raised,
    broad-exception-caught,
    consider-using-f-string,
    useless-parent-delegation
"""

[tool.pylint.master]
fail-under = 5.0
ignore = []

[tool.pylint.format]
max-line-length = 88
max-module-lines = 800

[[tool.mypy.overrides]]
module = [
    "test.*",
]
ignore_errors = true

[tool.sonar]
projectKey = "com.exasol:script-languages-package-management"
hostUrl = "https://sonarcloud.io"
organization = "exasol"
exclusions = "exasol/exaslpm/version.py"
